Moved code to own module for now
